campusgo_api/requirements.txt
flask
gunicorn
argon2-cffi
PyJWT
werkzeug
pymysql
python-dotenv
boto3
